Move shell/process into bootstrap/subprocess
Having shell/process and process ended up being very confusing. After
some digging the process package is used to create the bootstrap process
and then the shell/process is used as a simpler version for creating
subprocesses.

I also changed the name from shell.Process to bootstrap.Subprocess as I
think it helps to make the lineage of Agent -> Bootstrap Process ->
Bootstrap Subprocess clearer.
